Added files via upload
Added code into the Code folder for the Cartpole and Atari games. IPython notebooks and walkthroughs are available in the Tutorials folder. The presentation for Bayesian Optimization (BO) for Deep RL has been added to the Presentation folder. Finally, the references for all material and theory discussed, as well as for all deep learning sources and BO tutorials is available under References